版本V1.20171222   摄像表 表ID更改失败修复，修改底层关键工具类

+++ app/src/main/java/com/seck/hzy/lorameterapp/LoRaApp/utils/HzyUtils.java
@@ -349,7 +349,7 @@ public class HzyUtils {
         if (HzyUtils.isEmpty(str)) {
             return "0";
         } else {
-            return Integer.toHexString(Integer.parseInt(str));
+            return Long.toHexString(Long.parseLong(str));
         }
     }
